emontx_demo_9
[txshield_demo_9.tar.gz](https://community.openenergymonitor.org/uploads/default/original/2X/0/0c808295e0d0252e6ef5e04461627a9fff1ea094.gz)
[STM Deveplopment post #198](https://community.openenergymonitor.org/t/stm32-development/6815/198?u=pb66)

+++ emonTxshield/Inc/power.h
@@ -29,9 +29,9 @@ typedef enum channel_state_ {
 } channel_state_t;
 
 typedef struct power_stats_ {
-  float sigma_power;
-  float sigma_i_sq;
-  float sigma_v_sq;
+  int64_t sigma_power;
+  int64_t sigma_i_sq;
+  int64_t sigma_v_sq;
   int count;
   int sigma_i;
   int sigma_v;
